server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
boto3==1.35.0
pydantic==2.9.0
python-dotenv==1.0.1
mangum==0.18.0
